backend/perplexity_server/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
gunicorn 